+++ encoding/json/outbound.go
@@ -48,6 +48,10 @@ func New(c transport.Channel) Client {
 	return jsonClient{ch: c}
 }
 
+func init() {
+	yarpc.RegisterClientBuilder(New)
+}
+
 type jsonClient struct {
 	ch transport.Channel
 }