+++ examples/thrift/keyvalue/kv/yarpc/keyvalueclient/client.go
@@ -24,13 +24,13 @@
 package keyvalueclient
 
 import (
-	yarpc "go.uber.org/yarpc"
-	"golang.org/x/net/context"
 	"go.uber.org/thriftrw/protocol"
-	"go.uber.org/thriftrw/wire"
+	"golang.org/x/net/context"
 	"go.uber.org/yarpc/encoding/thrift"
 	"go.uber.org/yarpc/transport"
 	"go.uber.org/yarpc/examples/thrift/keyvalue/kv/service/keyvalue"
+	"go.uber.org/yarpc"
+	"go.uber.org/thriftrw/wire"
 )
 
 // Interface is a client for the KeyValue service.
@@ -58,6 +58,12 @@ func New(c transport.Channel, opts ...thrift.ClientOption) Interface {
 		Channel:  c,
 		Protocol: protocol.Binary,
 	}, opts...)}
+}
+
+func init() {
+	yarpc.RegisterClientBuilder(func(c transport.Channel) Interface {
+		return New(c)
+	})
 }
 
 type client struct{ c thrift.Client }